refactor: deduplicate heroItemsQuery from homescreen mobile/tv
