+++ components/home/Home.tv.tsx
@@ -34,6 +34,7 @@ import { TVHeroCarousel } from "@/components/home/TVHeroCarousel";
 import { Loader } from "@/components/Loader";
 import { useScaledTVTypography } from "@/constants/TVTypography";
 import useRouter from "@/hooks/useAppRouter";
+import { useHeroItems } from "@/hooks/useHeroItems";
 import { useNetworkStatus } from "@/hooks/useNetworkStatus";
 import { useRefreshLibraryOnFocus } from "@/hooks/useRefreshLibraryOnFocus";
 import { useInvalidatePlaybackProgressCache } from "@/hooks/useRevalidatePlaybackProgressCache";
@@ -220,56 +221,8 @@ export const Home = () => {
     refetchInterval: 60 * 1000,
   });
 
-  // Fetch hero items (Continue Watching + Next Up combined)
-  const { data: heroItems } = useQuery({
-    queryKey: ["home", "heroItems", user?.Id],
-    queryFn: async () => {
-      if (!api || !user?.Id) return [];
-
-      const [resumeResponse, nextUpResponse] = await Promise.all([
-        getItemsApi(api).getResumeItems({
-          userId: user.Id,
-          enableImageTypes: ["Primary", "Backdrop", "Thumb"],
-          includeItemTypes: ["Movie", "Series", "Episode"],
-          fields: ["Overview"],
-          startIndex: 0,
-          limit: 10,
-        }),
-        getTvShowsApi(api).getNextUp({
-          userId: user.Id,
-          startIndex: 0,
-          limit: 10,
-          fields: ["Overview"],
-          enableImageTypes: ["Primary", "Backdrop", "Thumb"],
-          enableResumable: false,
-        }),
-      ]);
-
-      const resumeItems = resumeResponse.data.Items || [];
-      const nextUpItems = nextUpResponse.data.Items || [];
-
-      // Combine, sort by recent activity, and dedupe
-      const combined = [...resumeItems, ...nextUpItems];
-      const sorted = combined.sort((a, b) => {
-        const dateA = a.UserData?.LastPlayedDate || a.DateCreated || "";
-        const dateB = b.UserData?.LastPlayedDate || b.DateCreated || "";
-        return new Date(dateB).getTime() - new Date(dateA).getTime();
-      });
-
-      const seen = new Set<string>();
-      const deduped: BaseItemDto[] = [];
-      for (const item of sorted) {
-        if (!item.Id || seen.has(item.Id)) continue;
-        seen.add(item.Id);
-        deduped.push(item);
-      }
-
-      return deduped.slice(0, 15);
-    },
-    enabled: !!api && !!user?.Id,
-    staleTime: 60 * 1000,
-    refetchInterval: 60 * 1000,
-  });
+  // Hero items (Continue Watching + Next Up combined), shared with Home.tsx
+  const { data: heroItems } = useHeroItems();
 
   useEffect(() => {
     updateTVDiscovery({
